chore(renovate): tweak schedule
If we don't specify a timeframe, Renovate could create multiple PRs
during the day once we merge the first PRs it creates, which can be
noisy.

According to
https://docs.renovatebot.com/configuration-options/#schedule, it doesn't
support exact time (probably because Renovate runs are not exactly
predictable), so setting "before 3am" and "before 5am" should give it
enough time to make PRs, while making it almost sure that we won't get
new PRs during the day after the first ones created are merged.